feat: add devops workflow mr step
